feat(core): Remove `provideZonelessChangeDetection` requirement when ZoneJS is not present
This change updates the internals to provide zoneless scheduling by
default when ZoneJS. This change means that for most applications,
simply omitting the ZoneJS polyfill will be enough to use zoneless
change detection.

Developers may still need to use `provideZonelessChangeDetection` in the providers if the application
exists alongside another one that uses ZoneJS (because ZoneJS would then
be present and the zone-based change detection scheduler would then be
selected).

This is one step towards zoneless by default, but does not yet go as far
as removing the `provideZoneChangeDetection` from the internals. That
would/will be larger breaking change since it will require existing
Zone-based applications to manually include `provideZoneChangeDetection`
to the application providers (and for all tests).

This approach works well because all existing Zone-based applications
_must_ have ZoneJS already included on the page for them to work. In
addition, this does not affect existing zoneless applications because
they all must already manually use `provideZonelessChangeDetection`
since the only scheduler provided by default prior to this change was
the zone-based one. Only "custom" zoneless (zoneless apps prior to real
zoneless support) may be affected.

BREAKING CHANGE: Applications that configure a custom zoneless implementation
now and _do not_ want the Angular zoneless scheduler to be enabled will
need to include `provideZoneChangeDetection` as well as a provider
for `NgZone` in the application providers:

```
bootstrapApplication(App, {providers: [
  provideZoneChangeDetection(),
  {provide: NgZone, useClass: MyNoopNgZone}
]});
```
