+++ packages/core/src/core_private_export.ts
@@ -37,11 +37,8 @@ export {
   defaultIterableDiffers as ɵdefaultIterableDiffers,
   defaultKeyValueDiffers as ɵdefaultKeyValueDiffers,
 } from './change_detection/change_detection';
-export {
-  internalProvideZoneChangeDetection as ɵinternalProvideZoneChangeDetection,
-  PROVIDED_NG_ZONE as ɵPROVIDED_NG_ZONE,
-} from './change_detection/scheduling/ng_zone_scheduling';
-export {ChangeDetectionSchedulerImpl as ɵChangeDetectionSchedulerImpl} from './change_detection/scheduling/zoneless_scheduling_impl';
+export {provideChangeDetectionScheduling as ɵprovideChangeDetectionScheduling} from './change_detection/scheduling/default_scheduling_provider';
+export {PROVIDED_NG_ZONE as ɵPROVIDED_NG_ZONE} from './change_detection/scheduling/ng_zone_scheduling';
 export {
   ChangeDetectionScheduler as ɵChangeDetectionScheduler,
   NotificationSource as ɵNotificationSource,
