+++ packages/core/test/bundling/animations-standalone/bundle.golden_symbols.json
@@ -524,6 +524,7 @@
       "getNearestLContainer",
       "getNextLContainer",
       "getNgDirectiveDef",
+      "getNgZone",
       "getNgZoneOptions",
       "getNodeInjectable",
       "getNullInjector",
@@ -591,6 +592,7 @@
       "internalCreateApplication",
       "internalImportProvidersFrom",
       "internalProvideZoneChangeDetection",
+      "internalProvideZonelessChangeDetection",
       "interpolateParams",
       "invalidCssUnitValue",
       "invalidDefinition",
@@ -747,6 +749,7 @@
       "profilerCallbacks",
       "projectNodes",
       "provideAnimations",
+      "provideChangeDetectionScheduling",
       "provideZonelessChangeDetection",
       "providerToFactory",
       "providerToRecord",
