+++ packages/core/test/bundling/defer/bundle.golden_symbols.json
@@ -474,6 +474,7 @@
       "getNearestLContainer",
       "getNextLContainer",
       "getNgDirectiveDef",
+      "getNgZone",
       "getNgZoneOptions",
       "getNodeInjectable",
       "getNullInjector",
@@ -543,6 +544,7 @@
       "internalCreateApplication",
       "internalImportProvidersFrom",
       "internalProvideZoneChangeDetection",
+      "internalProvideZonelessChangeDetection",
       "invokeAllTriggerCleanupFns",
       "invokeDirectivesHostBindings",
       "invokeHostBindingsInCreationMode",
@@ -649,6 +651,7 @@
       "profiler",
       "profilerCallbacks",
       "projectNodes",
+      "provideChangeDetectionScheduling",
       "provideZonelessChangeDetection",
       "providerToFactory",
       "providerToRecord",
