+++ packages/core/test/bundling/forms_reactive/bundle.golden_symbols.json
@@ -693,6 +693,7 @@
       "instructionState",
       "internalImportProvidersFrom",
       "internalProvideZoneChangeDetection",
+      "internalProvideZonelessChangeDetection",
       "invokeDirectivesHostBindings",
       "invokeHostBindingsInCreationMode",
       "isAbstractControlOptions",
@@ -875,6 +876,7 @@
       "profiler",
       "profilerCallbacks",
       "projectNodes",
+      "provideChangeDetectionScheduling",
       "providerToFactory",
       "providerToRecord",
       "providersResolver",
