+++ packages/core/test/bundling/forms_template_driven/bundle.golden_symbols.json
@@ -695,6 +695,7 @@
       "instructionState",
       "internalImportProvidersFrom",
       "internalProvideZoneChangeDetection",
+      "internalProvideZonelessChangeDetection",
       "invokeDirectivesHostBindings",
       "invokeHostBindingsInCreationMode",
       "isAngularZoneProperty",
@@ -874,6 +875,7 @@
       "profiler",
       "profilerCallbacks",
       "projectNodes",
+      "provideChangeDetectionScheduling",
       "providerToFactory",
       "providerToRecord",
       "providersResolver",
