+++ packages/core/test/bundling/hydration/bundle.golden_symbols.json
@@ -492,6 +492,7 @@
       "getNextLContainer",
       "getNgContainerSize",
       "getNgDirectiveDef",
+      "getNgZone",
       "getNgZoneOptions",
       "getNoOffsetIndex",
       "getNodeInjectable",
@@ -563,6 +564,7 @@
       "internalCreateApplication",
       "internalImportProvidersFrom",
       "internalProvideZoneChangeDetection",
+      "internalProvideZonelessChangeDetection",
       "invokeDirectivesHostBindings",
       "invokeHostBindingsInCreationMode",
       "isAngularZoneProperty",
@@ -688,6 +690,7 @@
       "profiler",
       "profilerCallbacks",
       "projectNodes",
+      "provideChangeDetectionScheduling",
       "provideClientHydration",
       "providerToFactory",
       "providerToRecord",
