+++ packages/core/test/bundling/router/bundle.golden_symbols.json
@@ -685,6 +685,7 @@
       "getNextLContainer",
       "getNgDirectiveDef",
       "getNgModuleDef",
+      "getNgZone",
       "getNgZoneOptions",
       "getNodeInjectable",
       "getNullInjector",
@@ -782,6 +783,7 @@
       "internalCreateApplication",
       "internalImportProvidersFrom",
       "internalProvideZoneChangeDetection",
+      "internalProvideZonelessChangeDetection",
       "interpolation1",
       "invokeDirectivesHostBindings",
       "invokeHostBindingsInCreationMode",
@@ -969,6 +971,7 @@
       "profiler",
       "profilerCallbacks",
       "projectNodes",
+      "provideChangeDetectionScheduling",
       "provideRouter",
       "provideZonelessChangeDetection",
       "providerToFactory",
