+++ packages/core/test/bundling/standalone_bootstrap/bundle.golden_symbols.json
@@ -393,6 +393,7 @@
       "getNearestLContainer",
       "getNextLContainer",
       "getNgDirectiveDef",
+      "getNgZone",
       "getNgZoneOptions",
       "getNodeInjectable",
       "getNullInjector",
@@ -451,6 +452,7 @@
       "internalCreateApplication",
       "internalImportProvidersFrom",
       "internalProvideZoneChangeDetection",
+      "internalProvideZonelessChangeDetection",
       "invokeDirectivesHostBindings",
       "invokeHostBindingsInCreationMode",
       "isAngularZoneProperty",
@@ -545,6 +547,7 @@
       "profiler",
       "profilerCallbacks",
       "projectNodes",
+      "provideChangeDetectionScheduling",
       "provideZonelessChangeDetection",
       "providerToFactory",
       "providerToRecord",
