+++ packages/platform-browser/testing/src/browser.ts
@@ -12,9 +12,7 @@ import {
   createPlatformFactory,
   NgModule,
   StaticProvider,
-  ɵinternalProvideZoneChangeDetection as internalProvideZoneChangeDetection,
-  ɵChangeDetectionScheduler as ChangeDetectionScheduler,
-  ɵChangeDetectionSchedulerImpl as ChangeDetectionSchedulerImpl,
+  ɵprovideChangeDetectionScheduling as provideChangeDetectionScheduling,
   PlatformRef,
 } from '@angular/core';
 import {TestComponentRenderer} from '@angular/core/testing';
@@ -38,8 +36,7 @@ export const platformBrowserTesting: (extraProviders?: StaticProvider[]) => Plat
   exports: [BrowserModule],
   providers: [
     {provide: APP_ID, useValue: 'a'},
-    internalProvideZoneChangeDetection({}),
-    {provide: ChangeDetectionScheduler, useExisting: ChangeDetectionSchedulerImpl},
+    provideChangeDetectionScheduling(),
     {provide: PlatformLocation, useClass: MockPlatformLocation},
     {provide: TestComponentRenderer, useClass: DOMTestComponentRenderer},
   ],
